requirements.txt
requests==2.28.2
python-dateutil==2.8.2
 